requirements.txt
playwright~=1.57.0
pytest~=6.2.4
pyyaml~=6.0.3
pytest-html
pytest-xdist
